Probably works now

+++ kadmium-dmx.Data/Types/Fixtures/IDMXChannelData.cs
@@ -7,8 +7,8 @@ namespace kadmium_dmx_data.Types.Fixtures
     public interface IDMXChannelData
     {
         string Name { get; set; }
-        int Min { get; set; }
-        int Max { get; set; }
-        int Address { get; set; }
+        byte Min { get; set; }
+        byte Max { get; set; }
+        ushort Address { get; set; }
     }
 }
